1. 取消object中毫无意义的奇偶哈希 2. 提升object标识为64位

+++ icom/object.h
@@ -3,7 +3,7 @@
 
 #include "compiler.h"
 
-typedef int objhld_t;
+typedef int64_t objhld_t;
 
 #define INVALID_OBJHLD		(~((objhld_t)(0)))
